fix: move api & remove ui
