+++ backend/courses/views.py
@@ -1,6 +1,5 @@
 import logging
 
-from assignments.models import Assignment
 from django.contrib.auth import get_user_model
 from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
@@ -529,79 +528,4 @@ class ClassStudentsView(APIView):  # GET /classes/{id}/students
                 error=str(e),
                 message="학생 등록 중 오류가 발생했습니다.",
                 status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            )
-
-
-class TeacherDashboardStatsView(APIView):
-    """교사 대시보드 통계 API"""
-
-    @swagger_auto_schema(
-        operation_id="교사 대시보드 통계",
-        operation_description="교사의 총 과제 수, 총 학생 수, 총 클래스 수를 조회합니다.",
-        manual_parameters=[
-            openapi.Parameter(
-                name="teacherId",
-                in_=openapi.IN_QUERY,
-                description="교사 ID",
-                type=openapi.TYPE_INTEGER,
-                required=True,
-            ),
-        ],
-        responses={
-            200: "Dashboard statistics",
-            400: "Invalid teacher ID",
-            404: "Teacher not found",
-        },
-    )
-    def get(self, request):
-        teacher_id = request.query_params.get("teacherId")
-
-        if not teacher_id:
-            return Response(
-                {"success": False, "message": "teacherId 파라미터가 필요합니다."},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        try:
-            # 교사 존재 확인
-            teacher = Account.objects.get(id=teacher_id, is_student=False)
-
-            # 총 과제 수 계산 (해당 교사가 생성한 과제)
-            total_assignments = Assignment.objects.filter(course_class__teacher=teacher).count()
-
-            # 총 학생 수 계산 (해당 교사의 클래스에 등록된 학생들)
-            total_students = (
-                Enrollment.objects.filter(course_class__teacher=teacher, status=Enrollment.Status.ENROLLED)
-                .values("student")
-                .distinct()
-                .count()
-            )
-
-            # 총 클래스 수 계산
-            total_classes = CourseClass.objects.filter(teacher=teacher).count()
-
-            return Response(
-                {
-                    "success": True,
-                    "data": {
-                        "total_assignments": total_assignments,
-                        "total_students": total_students,
-                        "total_classes": total_classes,
-                    },
-                    "message": "대시보드 통계 조회 성공",
-                    "error": None,
-                },
-                status=status.HTTP_200_OK,
-            )
-
-        except Account.DoesNotExist:
-            return Response(
-                {"success": False, "message": "해당 교사를 찾을 수 없습니다."},
-                status=status.HTTP_404_NOT_FOUND,
-            )
-        except Exception as e:
-            logger.error(f"[TeacherDashboardStatsView] {e}", exc_info=True)
-            return Response(
-                {"success": False, "error": str(e), "message": "통계 조회 중 오류가 발생했습니다."},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR,
             )
